[REFACTOR][PYTHON] Split backend loader from package init
