+++ python/tvm/backend/rocm/__init__.py
@@ -33,7 +33,7 @@ def _detect_target_from_device(dev):
 
 def register_backend():
     """Register ROCm-owned Python semantics."""
-    from tvm.backend._autoload_backends import _load_runtime_sidecar
+    from tvm.backend.loader import _load_runtime_sidecar
     from tvm.target.detect_target import register_device_target_detector
 
     _load_runtime_sidecar("rocm")
