+++ python/tvm/backend/vulkan/__init__.py
@@ -43,7 +43,7 @@ def _detect_target_from_device(dev):
 
 def register_backend():
     """Register Vulkan-owned Python semantics."""
-    from tvm.backend._autoload_backends import _load_runtime_sidecar
+    from tvm.backend.loader import _load_runtime_sidecar
     from tvm.target.detect_target import register_device_target_detector
 
     _load_runtime_sidecar("vulkan")
